fi: js files with eslint
